% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{depts}
\alias{depts}
\title{Information in Turkish about departments, faculties in Turkish universities.}
\format{
A data frame with 25652 rows and 7 variables:
\describe{
  \item{id}{University id}
  \item{year}{verinin yayinlandigi yil - year of the data is anounced}
  \item{type}{universite tipi (devlet, ozel) - type of university (state or private)}
  \item{program_code}{program kodu - official program id of the department}
  \item{university}{universite ismi - name of the university}
  \item{faculty}{fakulte ismi - name of the faculty}
  \item{department}{bolum ismi - name of the department}
}
}
\source{
{Yokatlas, YOK (Turkish Higher Education Council)}
}
\usage{
depts
}
\description{
A dataset containing type, faculty, deparment information in Turkish for 198 universties in Turkey .
}
\keyword{datasets}
