% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{depts_en}
\alias{depts_en}
\title{Information in English about departments, faculties in Turkish universities.}
\format{
A data frame with 25652 rows and 7 variables:
\describe{
  \item{id}{University id}
  \item{year}{year of the data is anounced}
  \item{type}{type of university (state or private)}
  \item{program_code}{official program id of the department}
  \item{university}{name of the university}
  \item{faculty}{name of the faculty}
  \item{department}{name of the department}
}
}
\source{
{Yokatlas, YOK (Turkish Higher Education Council)}
}
\usage{
depts_en
}
\description{
A dataset containing type, faculty, deparment information in English for 198 universties in Turkey .
}
\keyword{datasets}
