% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_score.R
\name{list_score}
\alias{list_score}
\title{List scores for universities, departments}
\usage{
list_score(
  region_names = "all",
  city_names = "all",
  university_names = "all",
  department_names,
  lang = "en",
  var_ids
)
}
\arguments{
\item{region_names}{The region names that you want to get university,department and faculty names for}

\item{city_names}{The city names that you want to get university,department and faculty names for}

\item{university_names}{The university names that you want to get university get university,department and faculty names for}

\item{department_names}{The department names that.}

\item{lang}{Language option. Default ="en" "en" for English, "tr" for Turkish}

\item{var_ids}{Statistics defined in Table 3.}
}
\value{
A data frame
}
\description{
List statistics universities, departments and faculties based on city, university and region filter.
}
\examples{
list_score(region_names = "all", city_names = "all", 
university_names = "all", department_names="all", 
lang = "en", var_ids=c("X10","X11"))
}
