% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_universities.R
\name{list_universities}
\alias{list_universities}
\title{List universities with aggregations}
\usage{
list_universities(region_names, city_names)
}
\arguments{
\item{region_names}{The region names that you want to get university names for}

\item{city_names}{The city names that you want to get university names for}
}
\value{
A data frame
}
\description{
List universities and types of them based on city and region filter
}
\examples{
uni_names <- list_universities(region_names = "AGEAN",city_names = c("IZMIR","MUGLA"))
}
