% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regions_cities}
\alias{regions_cities}
\title{Information in Turkish about universities in Turkey.}
\format{
A data frame with 198 rows and 6 variables:
\describe{
  \item{id}{University id}
  \item{university}{universite ismi, name of the university}
  \item{type}{universite tipi (devlet, ozel) - type of university (state or private)}
  \item{city}{universitenin yer aldigi sehir ismi - name of the city where the university is located}
  \item{region}{universitenin yer aldigi sehir ismi - name of the region where the university is located}
  \item{website}{universitenin web adresi - website url of the university}
}
}
\source{
{Yokatlas, YOK (Turkish Higher Education Council)}
}
\usage{
regions_cities
}
\description{
A dataset containing region, city, type and website information in Turkish for 198 universties in Turkey .
}
\keyword{datasets}
