% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regions_cities_en}
\alias{regions_cities_en}
\title{Information in English about universities in Turkey.}
\format{
A data frame with 198 rows and 6 variables:
\describe{
  \item{id}{University id}
  \item{university}{name of the university}
  \item{type}{type of university (state or private)}
  \item{city}{name of the city where the university is located}
  \item{region}{name of the region where the university is located}
  \item{website}{website url of the university}
}
}
\source{
{Yokatlas, YOK (Turkish Higher Education Council)}
}
\usage{
regions_cities_en
}
\description{
A dataset containing region, city, type and website information in English for 198 universties in Turkey .
}
\keyword{datasets}
