% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scores}
\alias{scores}
\title{Information in Turkish about departments, faculties in Turkish universities.}
\format{
A data frame with 25652 rows and 198 variables:
\describe{
  \item{id}{University id}
  \item{year}{verinin yayinlandigi yil - year of the data is anounced}
  \item{type}{universite tipi (devlet, ozel) - type of university (state or private)}
  \item{program_code}{program kodu - official program id of the department}
  \item{university}{universite ismi - name of the university}
  \item{faculty}{fakulte ismi - name of the faculty}
  \item{department}{bolum ismi - name of the department}
  \item{X10}{number of quota for the highest rank student in high school}
  \item{X100}{number of placed students from Bayburt}
  \item{X101}{number of placed students from Karaman}
  \item{X102}{number of placed students from Kirikkale}
  \item{X103}{number of placed students from Batman}
  \item{X104}{number of placed students from Sirnak}
  \item{X105}{number of placed students from Bartin}
  \item{X106}{number of placed students from Ardahan}
  \item{X107}{number of placed students from Igdir}
  \item{X108}{number of placed students from Yalova}
  \item{X109}{number of placed students from Karabuk}
  \item{X11}{number of placed to general quota}
  \item{X110}{number of placed students from Kilis}
  \item{X111}{number of placed students from Osmaniye}
  \item{X112}{number of placed students from Duzce}
  \item{X113}{number of students graduated from high school in same year}
  \item{X114}{number of students graduated from high school and not placed before}
  \item{X115}{number of students placed when they are students}
  \item{X116}{number of students graduated from university before}
  \item{X117}{number of student graduated from high school in same year}
  \item{X118}{number of student graduated from high school one year ago}
  \item{X119}{number of student graduated from high school two years ago}
  \item{X12}{number of students placed to the quota of highest rank student in high school}
  \item{X120}{number of student graduated from high school three years ago}
  \item{X121}{number of student graduated from high school four years ago}
  \item{X122}{lowest score of students placed quota of highest rank student in high school}
  \item{X123}{highest rank of students placed quota of highest rank student in high school}
  \item{X124}{average turkish test net}
  \item{X125}{average math test net}
  \item{X126}{average social studies test net}
  \item{X127}{average science test net}
  \item{X128}{average foreign language test net}
  \item{X129}{average ayt turkish language test net}
  \item{X13}{number of additional placement students}
  \item{X130}{average history test net}
  \item{X131}{avege ayt math test net}
  \item{X132}{average geography test net}
  \item{X133}{average obp score}
  \item{X134}{average tyt score}
  \item{X135}{lowest obp score}
  \item{X136}{lowest tyt score}
  \item{X137}{average rank of success}
  \item{X138}{lowest rank of success}
  \item{X139}{number of students who choice}
  \item{X14}{lowest score of students placed with coefficient 0.12}
  \item{X140}{average choice place}
  \item{X141}{number of choices in 1st place}
  \item{X142}{number of choices in 2nd place}
  \item{X143}{number of choices in 3rd place}
  \item{X144}{number of choices in 4th place}
  \item{X145}{number of choices in 5th place}
  \item{X146}{number of choices in 6th place}
  \item{X147}{number of choices in 7th place}
  \item{X148}{number of choices in 8th place}
  \item{X149}{number of choices in 9th place}
  \item{X15}{lowest rank of students placed with coefficient 0.12}
  \item{X150}{number of choices in 10th and more place}
  \item{X151}{number of students placed in 1st choice}
  \item{X152}{number of students placed in 2nd choice}
  \item{X153}{number of students placed in 3rd choice}
  \item{X154}{number of students placed in 4th choice}
  \item{X155}{number of students placed in 5th choice}
  \item{X156}{number of students placed in 6th choice}
  \item{X157}{number of students placed in 7th choice}
  \item{X158}{number of students placed in 8th choice}
  \item{X159}{number of students placed in 9th choice}
  \item{X16}{highest score of placed students}
  \item{X160}{number of students placed in 10th choice}
  \item{X161}{number of students placed in 11th choice}
  \item{X162}{number of students placed in 12th choice}
  \item{X163}{number of students placed in 13th choice}
  \item{X164}{number of students placed in 14th choice}
  \item{X165}{number of students placed in 15th choice}
  \item{X166}{number of students placed in 16th choice}
  \item{X167}{number of students placed in 17th choice}
  \item{X168}{number of students placed in 18th choice}
  \item{X169}{number of students placed in 19th choice}
  \item{X17}{highest rank of placed students}
  \item{X170}{number of students placed in 20th choice}
  \item{X171}{number of students placed in 21th choice}
  \item{X172}{number of students placed in 22th choice}
  \item{X173}{number of students placed in 23th choice}
  \item{X174}{number of students placed in 24th choice}
  \item{X175}{total used choices of students}
  \item{X176}{used number of choices}
  \item{X177}{unused number of choices}
  \item{X178}{average used choices of students}
  \item{X179}{total used choices of students}
  \item{X18}{the number of students who placed in the previous year and replaced with the deduction of obp score}
  \item{X180}{number of choices for private universities}
  \item{X181}{number of choices for Cyprus universities}
  \item{X182}{number of choices for abroad universities}
  \item{X183}{number of choices for same department/program}
  \item{X184}{number of choices for different department/program}
  \item{X185}{number of choices for Cyprus department/program}
  \item{X186}{number of choices for programs}
  \item{X187}{number of choices for abroad department/program}
  \item{X188}{number of full professor}
  \item{X189}{number of associate professor}
  \item{X19}{average obp score}
  \item{X190}{number of assistant professor}
  \item{X191}{number of registered female students}
  \item{X192}{number of registered male students}
  \item{X193}{number of male graduates}
  \item{X194}{number of female graduates}
  \item{X195}{number of outgoing exchange students}
  \item{X196}{number of incoming exchange students}
  \item{X2}{program code}
  \item{X20}{average high school graduation score}
  \item{X21}{number of placed female students}
  \item{X22}{number of placed male students}
  \item{X23}{number of placed students from same city}
  \item{X24}{number of placed students from different city}
  \item{X25}{number of placed students from mediterranean region}
  \item{X26}{number of placed students from eastern anatolia region}
  \item{X27}{number of placed students from aegean region}
  \item{X28}{number of placed students from southeaster anatolia region}
  \item{X29}{number of placed students from central anatolia region}
  \item{X30}{number of placed students from Black sea region}
  \item{X31}{number of placed students from Marmara region}
  \item{X32}{number of placed students from Adana}
  \item{X33}{number of placed students from Adiyaman}
  \item{X34}{number of placed students from Afyon}
  \item{X35}{number of placed students from Agri}
  \item{X36}{number of placed students from Amasya}
  \item{X37}{number of placed students from Ankara}
  \item{X38}{number of placed students from Antalya}
  \item{X39}{number of placed students from Artvin}
  \item{X40}{number of placed students from Aydin}
  \item{X41}{number of placed students from Balikesir}
  \item{X42}{number of placed students from Bilecik}
  \item{X43}{number of placed students from Bingol}
  \item{X44}{number of placed students from Bitlis}
  \item{X45}{number of placed students from Bolu}
  \item{X46}{number of placed students from Burdur}
  \item{X47}{number of placed students from Bursa}
  \item{X48}{number of placed students from Canakkale}
  \item{X49}{number of placed students from Cankiri}
  \item{X50}{number of placed students from Corum}
  \item{X51}{number of placed students from Denizli}
  \item{X52}{number of placed students from Diyarbakir}
  \item{X53}{number of placed students from Edirne}
  \item{X54}{number of placed students from Elazig}
  \item{X55}{number of placed students from Erzincan}
  \item{X56}{number of placed students from Erzurum}
  \item{X57}{number of placed students from Eskisehir}
  \item{X58}{number of placed students from Gaziantep}
  \item{X59}{number of placed students from Giresun}
  \item{X6}{Type of score}
  \item{X60}{number of placed students from Gumushane}
  \item{X61}{number of placed students from Hakkari}
  \item{X62}{number of placed students from Hatay}
  \item{X63}{number of placed students from Isparta}
  \item{X64}{number of placed students from Mersin}
  \item{X65}{number of placed students from Istanbul}
  \item{X66}{number of placed students from Izmir}
  \item{X67}{number of placed students from Kars}
  \item{X68}{number of placed students from Kastamonu}
  \item{X69}{number of placed students from Kayseri}
  \item{X7}{type of scholarship}
  \item{X70}{number of placed students from Kirklareli}
  \item{X71}{number of placed students from Kirsehir}
  \item{X72}{number of placed students from Kocaeli}
  \item{X73}{number of placed students from Konya}
  \item{X74}{number of placed students from Kutahya}
  \item{X75}{number of placed students from Malatya}
  \item{X76}{number of placed students from Manisa}
  \item{X77}{number of placed students from Kahramanmaras}
  \item{X78}{number of placed students from Mardin}
  \item{X79}{number of placed students from Mugla}
  \item{X80}{number of placed students from Mus}
  \item{X81}{number of placed students from Nevsehir}
  \item{X82}{number of placed students from Nigde}
  \item{X83}{number of placed students from Ordu}
  \item{X84}{number of placed students from Rize}
  \item{X85}{number of placed students from Sakarya}
  \item{X86}{number of placed students from Samsun}
  \item{X87}{number of placed students from Siirt}
  \item{X88}{number of placed students from Sinop}
  \item{X89}{number of placed students from Sivas}
  \item{X9}{general contingency}
  \item{X90}{number of placed students from Tekirdag}
  \item{X91}{number of placed students from Tokat}
  \item{X92}{number of placed students from Trabzon}
  \item{X93}{number of placed students from Tunceli}
  \item{X94}{number of placed students from Sanliurfa}
  \item{X95}{number of placed students from Usak}
  \item{X96}{number of placed students from Van}
  \item{X97}{number of placed students from Yozgat}
  \item{X98}{number of placed students from Zonguldak}
  \item{X99}{number of placed students from Aksaray}
}
}
\source{
{Yokatlas, YOK (Turkish Higher Education Council)}
}
\usage{
scores
}
\description{
A dataset containing type, faculty, deparment information and other statistcs in Turkish for 198 universties in Turkey .
}
\keyword{datasets}
